+++ code/OptimizeMeshes.cpp
@@ -145,7 +145,7 @@ void OptimizeMeshesProcess::Execute( aiScene* pScene)
     meshes.resize( 0 );
     ai_assert(output.size() <= num_old);
 
-    mScene->mNumMeshes = output.size();
+    mScene->mNumMeshes = static_cast<unsigned int>(output.size());
     std::copy(output.begin(),output.end(),mScene->mMeshes);
 
     if (output.size() != num_old) {
@@ -199,7 +199,7 @@ void OptimizeMeshesProcess::ProcessNode( aiNode* pNode)
             } else {
                 output.push_back(mScene->mMeshes[im]);
             }
-            im = output.size()-1;
+            im = static_cast<unsigned int>(output.size()-1);
         }
     }
